Fan mode done. Fixed a mistake created during yesterday's code refactoring
New Blueprint Class: BP_FanCollectable
Collecting it will enter the Fan mode for [FanDuration]s (currently 5s).
During the Fan mode, players just need to hold the height of the pump handles to choose the floating force. The lower the pump handles are, the bigger the floating force is.
For now, it is still represented by Balloon sizes. If we still want the balloons' models turn into fans, we'll need @Artists , @Tech Artists , and @Audio/UI designer . @Producer
For now, the outfit of the collectable is just a star without any material, we'll need @Artists and @Tech Artists to design and implement the model and VFX of the collectable.