Core mechanics for sticky obstacles done
Core mechanics for Sticky Obstacles done!
New Blueprint Classes:
I have made a new child class BP_Obstacle_Sticky derived from the BP_Obstacle. Then a new child class BP_Obstacle_Sticky_Egg derived from the BP_Obstacle_Sticky.
A BP_Obstacle_Sticky will be destroyed when overlapping with a BP_Ship. Then it will detect the overlapped component on the BP_Ship to see whether it's balloon1, balloon2,  or a ship. Then trigger the different events from BP_Ship.
For now, the sticky obstacle's effect for balloons is similar to the spikes, which is disabling the balloon it touched for a few seconds. And the sticky obstacle's effect for the ship body is making it much heavier for a few seconds.