Added a deviation variable
Made the deviation (to reduce big jumps caused by sensor instability) to a variable in the DistanceSensor_Blueprint. It was hard-coded 10. And I changed it to 15 to reduce the circumstances that controller got temporary disabled while the player moves too fast. But this change might introduce more jumpings, the perfect solution to this would only be getting better sensors.